WEA-22 Refactor DailyData into its own class
