+++ app/src/main/java/com/ayush/weatherapp/home/TabPagerAdapter.java
@@ -4,7 +4,7 @@ import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.DailyForecast;
+import com.ayush.weatherapp.retrofit.weatherApi.pojo.DailyData;
 import com.ayush.weatherapp.retrofit.weatherApi.pojo.HourlyForecast;
 import java.util.List;
 
@@ -12,7 +12,7 @@ public class TabPagerAdapter extends FragmentPagerAdapter {
   private final static int NUMBER_OF_TABS = 2;
   private final static String TAB_TITLES[] = new String[] { "Daily", "Hourly" };
   private List<HourlyForecast.HourlyData> hourlyData;
-  private List<DailyForecast.DailyData> dailyData;
+  private List<DailyData> dailyData;
 
   TabPagerAdapter(FragmentManager fm) {
     super(fm);
@@ -35,7 +35,7 @@ public class TabPagerAdapter extends FragmentPagerAdapter {
     return TAB_TITLES[position];
   }
 
-  public void setDailyForecastData(List<DailyForecast.DailyData> dailyData) {
+  public void setDailyForecastData(List<DailyData> dailyData) {
     this.dailyData = dailyData;
   }
